added an interactive option
*changed the script to allow for the optional inclusion of a bash script
so that if you don't include one, you can interact with the terminal in
the container.

NOTE: this feature isn't perfect yet. If you run start_script.sh on it's
own it works great but if you run it through the make_test.py it's
really not perfect. I think this is because the python script is using a
loop and waiting for output from the call to the start_contianers.sh
script to add it to a log file. It still works and it interactive but
you can't see what you're typing until after you hit enter which isn't
desireable.

Signed-off-by: Jeremy Spewock <jspewock@iol.unh.edu>