Updated README and dockerfiles
*Updated README to include and exmaple script for building a Linux
kernel and also explained in the "running test scripts" section that it
is important that you are within the same directory as your scripts

*Added the wget package to dockerfiles as this is something that I
believe is used fairly often for downloading things off the internet
from the command line

NOTE: container images with the new wget dependancy were still building
at the time of this commit but adding this extra package shouldn't cause
issue. If there are any problems after testing another patch will follow to
correct the issue.

Signed-off-by: Jeremy Spewock <jspewock@iol.unh.edu>